requirements.txt
Adafruit-Blinka==2.4.0
adafruit-circuitpython-ads1x15==2.1.1
adafruit-circuitpython-busdevice==4.0.0
Adafruit-PlatformDetect==1.3.3
Adafruit-PureIO==0.2.3
Click==7.0
Flask==1.1.1
Flask-SocketIO==4.2.1
gas-detection==1.0.1
gevent==1.4.0
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
python-engineio==3.9.3
python-socketio==4.3.1
rpi-ws281x==4.2.2
RPi.GPIO==0.7.0
six==1.12.0
spidev==3.4
sysv-ipc==1.0.0
Werkzeug==0.15.6
